Renamed new_logger to logger
